LayoutTests/media/media-session/actionHandler-no-document-leak-expected.txt

Tests that page installed actionHandlers do not leak documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isDocumentAlive(frameDocumentID) is true
PASS The iframe document didn't leak.
PASS successfullyParsed is true

TEST COMPLETE
